Add validation, feedback view, repository and navigation
Implemented a feedback view to enhance user interaction and added a repository to the model for in-memory storage. Included a validation statement to prevent duplicate course applications based on email. Updated the layout with a navigation bar and modified the URL links in the layout. Additionally, added a table in Course/Index.cshtml to display the list of attendees for each course.